--- tenant/aspnet/Xyz.Tenant.Api/WeatherForecast.cs
@@ -1,16 +1,0 @@
-//TODO: Delete this file
-using System;
-
-namespace Xyz.Tenant.Api
-{
-    public class WeatherForecast
-    {
-        public DateTime Date { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-
-        public string Summary { get; set; }
-    }
-}
