--- tenant/aspnet/Xyz.Tenant.DataAccess/Mapper.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace Xyz.Tenant.DataAccess
-{
-    public class Mapper
-    {
-    }
-}
