--- tenant/aspnet/Xyz.Tenant.Lib/Xyz.Tenant.Lib.csproj
@@ -1,14 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="BusinessLogic\" />
-    <Folder Include="Exceptions\" />
-    <Folder Include="Interface\" />
-    <Folder Include="Models\" />
-  </ItemGroup>
-
-</Project>
