--- tenant/aspnet/Xyz.Tenant.sln
@@ -1,43 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29509.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Xyz.Tenant.Api", "Xyz.Tenant.Api\Xyz.Tenant.Api.csproj", "{F9958D5B-5B4D-44C0-92DB-F141B73A657A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xyz.Tenant.DataAccess", "Xyz.Tenant.DataAccess\Xyz.Tenant.DataAccess.csproj", "{75B7A25A-64EA-4254-B8A3-947A78FAD088}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xyz.Tenant.Lib", "Xyz.Tenant.Lib\Xyz.Tenant.Lib.csproj", "{850CCDCB-BF39-42CE-9237-EE1B47C9F3FE}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Xyz.Tenant.Tests", "Xyz.Tenant.Tests\Xyz.Tenant.Tests.csproj", "{8F918DDB-98FF-4A20-A680-5FDA1E845DDA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F9958D5B-5B4D-44C0-92DB-F141B73A657A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9958D5B-5B4D-44C0-92DB-F141B73A657A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9958D5B-5B4D-44C0-92DB-F141B73A657A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9958D5B-5B4D-44C0-92DB-F141B73A657A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{75B7A25A-64EA-4254-B8A3-947A78FAD088}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{75B7A25A-64EA-4254-B8A3-947A78FAD088}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{75B7A25A-64EA-4254-B8A3-947A78FAD088}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{75B7A25A-64EA-4254-B8A3-947A78FAD088}.Release|Any CPU.Build.0 = Release|Any CPU
-		{850CCDCB-BF39-42CE-9237-EE1B47C9F3FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{850CCDCB-BF39-42CE-9237-EE1B47C9F3FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{850CCDCB-BF39-42CE-9237-EE1B47C9F3FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{850CCDCB-BF39-42CE-9237-EE1B47C9F3FE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8F918DDB-98FF-4A20-A680-5FDA1E845DDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8F918DDB-98FF-4A20-A680-5FDA1E845DDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8F918DDB-98FF-4A20-A680-5FDA1E845DDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8F918DDB-98FF-4A20-A680-5FDA1E845DDA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {46F4CFF1-737A-4C9B-9474-F60EB08D29C5}
-	EndGlobalSection
-EndGlobal
