+++ tenant/src/Revature.Tenant.DataAccess/Mapper.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Revature.Tenant.DataAccess
+{
+    public class Mapper
+    {
+    }
+}
